+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 All notables changes between versions are documented in this file.
 
-## v0.2.2 - ????-??-??
+## v0.3.0 - ????-??-??
 
 ### Changes
 
 - TI-99/4A and CoCo2 configurations have been added to the [third-party configs](config/third-party/README.md) directory.  Thanks to [@keronian](https://github.com/keronian) for contributing these.
+- Stored off image files used to create the firmware in `output/images/`.  This allows post build inspection of the images used.  It also enables additional tests - `sdrr-info` can be used as an additional check (along with `test`), to ensure the images in the firmware are correct, and validate the behaviour of `sdrr-info` and `test` to be compared.
+- Substantial refactoring of `sdrr-gen`, to make it more maintainable.
+- Substantial refactoring of `sdrr-fw-parser` in order to make it suitable for airfrog integration.
 
 ### Fixes